requirements.txt
fastapi
uvicorn
jinja2
httpx
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
itsdangerous
